+++ app/edge/devices/components/EdgeDevicesDatatable/EdgeDevicesDatatable.tsx
@@ -9,8 +9,8 @@ import {
 } from 'react-table';
 import { useRowSelectColumn } from '@lineup-lite/hooks';
 
-import { Environment } from "@/portainer/environments/types";
-import { PaginationControls } from '@/portainer/components/pagination-controls';
+import { Environment } from "Portainer/environments/types";
+import { PaginationControls } from 'Portainer/components/pagination-controls';
 import {
   Table,
   TableActions,
@@ -20,23 +20,23 @@ import {
   TableSettingsMenu,
   TableTitle,
   TableTitleActions,
-} from '@/portainer/components/datatables/components';
-import { multiple } from '@/portainer/components/datatables/components/filter-types';
-import { useTableSettings } from '@/portainer/components/datatables/components/useTableSettings';
-import { ColumnVisibilityMenu } from '@/portainer/components/datatables/components/ColumnVisibilityMenu';
-import { useRepeater } from '@/portainer/components/datatables/components/useRepeater';
-import { useDebounce } from '@/portainer/hooks/useDebounce';
+} from 'Portainer/components/datatables/components';
+import { multiple } from 'Portainer/components/datatables/components/filter-types';
+import { useTableSettings } from 'Portainer/components/datatables/components/useTableSettings';
+import { ColumnVisibilityMenu } from 'Portainer/components/datatables/components/ColumnVisibilityMenu';
+import { useRepeater } from 'Portainer/components/datatables/components/useRepeater';
+import { useDebounce } from 'Portainer/hooks/useDebounce';
 import {
   useSearchBarContext,
   SearchBar,
-} from '@/portainer/components/datatables/components/SearchBar';
-import { useRowSelect } from '@/portainer/components/datatables/components/useRowSelect';
-import { Checkbox } from '@/portainer/components/form-components/Checkbox';
-import { TableFooter } from '@/portainer/components/datatables/components/TableFooter';
-import { SelectedRowsCount } from '@/portainer/components/datatables/components/SelectedRowsCount';
-import {EdgeDeviceTableSettings} from "@/edge/components/EdgeDevicesDatatable/types";
-import {EdgeDevicesDatatableSettings} from "@/edge/components/EdgeDevicesDatatable/EdgeDevicesDatatableSettings";
-import {EdgeDevicesDatatableActions} from "@/edge/components/EdgeDevicesDatatable/EdgeDevicesDatatableActions";
+} from 'Portainer/components/datatables/components/SearchBar';
+import { useRowSelect } from 'Portainer/components/datatables/components/useRowSelect';
+import { Checkbox } from 'Portainer/components/form-components/Checkbox';
+import { TableFooter } from 'Portainer/components/datatables/components/TableFooter';
+import { SelectedRowsCount } from 'Portainer/components/datatables/components/SelectedRowsCount';
+import {EdgeDeviceTableSettings} from "@/edge/devices/types";
+import {EdgeDevicesDatatableSettings} from "@/edge/devices/components/EdgeDevicesDatatable/EdgeDevicesDatatableSettings";
+import {EdgeDevicesDatatableActions} from "@/edge/devices/components/EdgeDevicesDatatable/EdgeDevicesDatatableActions";
 
 import { useColumns } from './columns';
