+++ app/edge/devices/components/EdgeDevicesDatatable/EdgeDevicesDatatableSettings.tsx
@@ -1,6 +1,6 @@
-import { TableSettingsMenuAutoRefresh } from '@/portainer/components/datatables/components/TableSettingsMenuAutoRefresh';
-import { useTableSettings } from '@/portainer/components/datatables/components/useTableSettings';
-import { EdgeDeviceTableSettings } from "@/edge/components/EdgeDevicesDatatable/types";
+import { TableSettingsMenuAutoRefresh } from 'Portainer/components/datatables/components/TableSettingsMenuAutoRefresh';
+import { useTableSettings } from 'Portainer/components/datatables/components/useTableSettings';
+import { EdgeDeviceTableSettings } from "@/edge/devices/types";
 
 export function EdgeDevicesDatatableSettings() {
   const { settings, setTableSettings } = useTableSettings<
