+++ app/edge/devices/components/EdgeDevicesDatatable/columns/index.tsx
