+++ app/edge/devices/components/EdgeDevicesDatatable/columns/state.tsx
@@ -1,7 +1,7 @@
 import { Column } from 'react-table';
 
-import { Environment } from "@/portainer/environments/types";
-import { DefaultFilter } from '@/portainer/components/datatables/components/Filter';
+import { Environment } from "Portainer/environments/types";
+import { DefaultFilter } from 'Portainer/components/datatables/components/Filter';
 
 export const state: Column<Environment> = {
   Header: 'State',
