+++ app/edge/devices/types.ts
